# Memorize the bit depth for drawable.