For context, please see https://github.com/McCLIM/McCLIM/issues/637 -- the
performance of some implementations of X (in this case, XQuartz) justifies this
memoization.