@kc- 2019-08-18
@scymtym I put in the aesthetic changes we spoke of. Unfortunately, `def-clx-class` uses structures, so I couldn't use a `:documentation` slot option, which I would have preferred.