@JMC-design 2019-08-18
I'm not sure I see the point of this commit.  WITH-GEOMETRY keeps track of drawable-depth.  Was there an actual problem with the existing cache?  Requests like that should get bundled properly into the queue and not cause extra requests. If the problem lies there it needs to be corrected for everything else that uses it.  The mcclim issue does not clarify this commit, why would drawable-depth even be called that often?