@kc- 2019-08-18
@JMC-design thank you for your comment. I absolutely do not pretend to know my way around CLX very well, but I can expand the context on the McCLIM issue.
Here's a [profile using XQuartz](https://gist.github.com/kc-/7e83e1b80e86354c5523af2b6367700b#file-gistfile1-txt-L59) ,
and [on Linux/X11](https://gist.github.com/scymtym/d353afc63eed0200ceb97beb6806abcd#file-profile-lisp-L13), you can tell the factor of around 20 between the two X implementations.

> why would drawable-depth even be called that often

For the pattern drawing (the test code in https://github.com/McCLIM/McCLIM/issues/637), the call chain goes roughly like this:

[draw-pattern*]
-> [draw-rectangle*]
->    [medium-draw-rectangle*](https://github.com/McCLIM/McCLIM/blob/master/Backends/CLX/medium.lisp#L673)
->         [with-clx-graphics](https://github.com/McCLIM/McCLIM/blob/master/Backends/CLX/medium.lisp#L484)
->              [design-gcontext](https://github.com/McCLIM/McCLIM/blob/master/Backends/CLX/medium.lisp#L435)
->                 [compute-rgb-image](https://github.com/McCLIM/McCLIM/blob/master/Backends/CLX/medium.lisp#L409), in which the depth is queried.

If your point is "don't do that, then!", fair enough -- maybe you can propose a superior fix for the performance issue with this, other than caching that value in user code.