@JMC-design 2019-08-18
Dont do that. Seriously though the gc code there doesn't seem unified and you seem to be wasting cycles checking and creating new gcs.  
That profile doesn't real show why its caught up in draw rectangles. 
My concern here is a bandaid solution possibly covering up the real problem just to get an implementation of clim to work performantly. The clx attribute and context cache shouldn't be inflating times x20, nor should we need two things accomplishing the same goal.