@kc- 2019-08-19
The factor of 20 is the macOS' XQuartz tax, I don't know why it is that much slower than on linux.

Maybe we can discuss this first: do you agree that asking CLX for an immutable value (the drawable depth) could do without the network connection every time? In that context, I don't think this commit is the wrong thing to change.

Maybe others can weigh in with their opinions on this.