@JMC-design 2019-08-19
The manual mentions reusing gcontexts is usually more efficient.

Your question is the wrong one, considering that drawable-depth should already be cached and shouldn't require a network connection. 

So like I said in the beginning, if the problem lies with the existing cache then that needs to be investigated since it caches all window/drawable attributes, and not covered up by keeping track of drawable depth in a second manner to fix performance of one product, on one operating system.

Perhaps the issue is competition for the main thread on macos since certain things have to be run on it?

p.s. that profile still doesn't show what you are saying it shows, perhaps there was a problem in the conversion process, quite wasteful, of turning your pattern into something to be displayed.  I mean, has anybody else had any of the problems you mentioned? Besides people who were running the clos version?