@kc- 2019-08-19
Which question is the wrong one, "What do you suggest?" or "Where can I find reference material?" ?

I already said that I'm not intimately familiar with CLX. You're uncomfortable with this patch, and I acknowledged that -- maybe at this point you could be a little more forthcoming with a characterization of what The Right Way would be, please.

Is `xlib:drawable-depth` supposed to access this cache you mentioned, or is it the wrong accessor for that purpose? Is *WINDOW-ATTRIBUTES* the cache you mean, and should user code like mcclim access that?

>that profile still doesn't show what you are saying it shows

I don't understand what you mean. I didn't put `xlib:drawable-depth` in the profile list because at that point I was still looking for the cause. Later, I localized it more narrowly -- to the call to `xlib:drawable-depth` inside of the macroexpansion of [code you probably don't care about]-- and after this patch the performance is back to what it should have been. That is what I am saying it shows, even without an explicit mention of xlib in the profile table.
At what point do we disagree?