@JMC-design 2019-08-19
drawable-depth should already be cached and shouldn't require a network connection