@JMC-design 2019-08-20
Just to make sure you understand the last comment.  It is not an agreement with you, it is the exact same words I said before because I dislike repeating myself, but here I go.

WITH-GEOMETRY inside DRAWABLE-DEPTH already caches the depth information.

If this is affecting times it needs to be invesigated and not covered up with a bandaid.