@kc- 2019-08-21
I repeated myself verbatim because you did the same thing. I tried asking for relevant information but you don't seem to want to follow up, which is fine of course.