@JMC-design 2019-08-21
How much more relevant do you need than WITH-GEOMETRY already caches DRAWABLE-DEPTH, you shouldn't need to cache it again!  Are you understanding you are duplicating behaviour to compensate for something only you have seen and haven't provided proof of and might not even be the fault of CLX!