@dkochmanski 2019-08-21
If I understand correctly `with-state` caches results until we leave its dynamic extent while this cache change makes it cache the depth inside drawable object (so it is persistent across different calls). It is not very obvious from the client perspective, i.e McCLIM creates drawables which are reused and with-state can't be instantiated globally, so there is certainly something into it.

That said it could be easily mitigated with the plist extension hook from the client code (I.e McCLIM's backend specification). Something in line of:

```
(defun cached-drawable-depth (drawable)
  (or (getf (xlib:drawable-plist drawable) :cached-depth)
        (setf (getf (xlib:drawable-plist drawable) :cached-depth)
                 (xlib:drawable-depth drawable))))
```

@JMC-design all dissatisfaction should be directed towards me, I gave a green light for this PR without knowing about `with-state` (I should have known better).

OTOH please note, that `with-state` is not feasible for McCLIM model and given that drawable-depth query is very slow on OSX it is a valid solution. I rather agree though that such caching should be done on client side; but I don't have very strong opinion about that.

@kc- do you agree with reverting this change and adding this "cache" on McCLIM side?