@JMC-design 2019-08-21
I do not know how mcclim deals with events, multiple clients,etc.., but if you use the windows plist you need to make sure you never try to access the plist on a window returned by an event.

My main concern is still what is causing this slowdown? 1 call to drawable depth should not inflate times like that. You could add an extra 20 network calls per cycle and it wouldn't come close to being that slow. So what is going on here? 

another possibility is rehashing of the xid hashtable.  I have also noticed some issues with deletion of xids sometimes not happening and then getting huge slowdowns when instantiating thousands of items. It's on my todo list when I start dealing with thousands of objects not in testing.