@kc- 2019-08-21
Let's start over, okay?

> My main concern is still what is causing this slowdown? 

I don't know.
What I can offer as a conversation piece is a more explicit comparison of the timing without and with this changeset (these are the times profiled for running `run-drawtest` from the McCLIM issue report)

Without:
```
  seconds  |     gc     |   consed   |  calls |  sec/call  |  name  
---------------------------------------------------------
    16.257 |      0.000 |    259,600 |  1,369 |   0.011875 | XLIB:DRAWABLE-DEPTH
     0.116 |      0.000 |  3,671,472 |  1,352 |   0.000086 | CLIM-CLX::DESIGN-GCONTEXT
     0.089 |      0.000 |  7,953,696 |  2,035 |   0.000044 | MEDIUM-DRAW-RECTANGLE*
     0.016 |      0.000 |  3,076,896 |  9,665 |   0.000002 | TRANSFORM-REGION
     0.009 |      0.000 |    754,288 |  1,371 |   0.000007 | CLIM-CLX::MEDIUM-GCONTEXT
     0.009 |      0.000 |     32,768 |  1,535 |   0.000006 | COMPOSE-TRANSFORMATIONS
     0.008 |      0.000 |    163,680 |  2,704 |   0.000003 | XLIB:CREATE-IMAGE
     0.007 |      0.000 |          0 |  6,137 |   0.000001 | MEDIUM-INK
     0.007 |      0.000 |          0 |  4,056 |   0.000002 | XLIB:FREE-GCONTEXT
     0.003 |      0.000 |          0 |  2,880 |   0.000001 | SHEET-NATIVE-TRANSFORMATION
     0.003 |      0.000 |          0 |  4,775 |   0.000001 | MEDIUM-LINE-STYLE
     0.002 |      0.000 |    524,528 |    676 |   0.000003 | DRAW-PATTERN*
     0.002 |      0.000 |          0 |    680 |   0.000003 | DRAW-RECTANGLE*
     0.001 |      0.000 |          0 |     16 |   0.000055 | MEDIUM-DRAW-POLYGON*
     0.000 |      0.000 |          0 |    980 |   0.000000 | RECTANGLE-EDGES*
     0.000 |      0.000 |          0 |    676 |   0.000000 | DRAW-DESIGN
---------------------------------------------------------
    16.529 |      0.000 | 16,436,928 | 40,907 |            | Total
```

With:
```
  seconds  |     gc     |   consed   |  calls |  sec/call  |  name  
---------------------------------------------------------
     0.053 |      0.000 |  4,453,360 |  1,352 |   0.000039 | CLIM-CLX::DESIGN-GCONTEXT
     0.035 |      0.000 |  8,148,208 |  2,035 |   0.000017 | MEDIUM-DRAW-RECTANGLE*
     0.021 |      0.000 |  1,932,448 |  9,665 |   0.000002 | TRANSFORM-REGION
     0.018 |      0.000 |          0 |  1,369 |   0.000013 | XLIB:DRAWABLE-DEPTH
     0.009 |      0.000 |     32,768 |  2,704 |   0.000003 | XLIB:CREATE-IMAGE
     0.005 |      0.000 |          0 |  4,775 |   0.000001 | MEDIUM-LINE-STYLE
     0.004 |      0.000 |          0 |  4,056 |   0.000001 | XLIB:FREE-GCONTEXT
     0.004 |      0.000 |     32,752 |  1,555 |   0.000002 | COMPOSE-TRANSFORMATIONS
     0.003 |      0.000 |          0 |  6,137 |   0.000001 | MEDIUM-INK
     0.001 |      0.000 |          0 |  3,183 |   0.000000 | SHEET-NATIVE-TRANSFORMATION
     0.001 |      0.000 |     32,768 |     16 |   0.000055 | MEDIUM-DRAW-POLYGON*
     0.000 |      0.000 |          0 |  1,030 |   0.000000 | RECTANGLE-EDGES*
     0.000 |      0.000 |          0 |    676 |   0.000000 | DRAW-DESIGN
     0.000 |      0.000 |          0 |    680 |   0.000000 | DRAW-RECTANGLE*
     0.000 |      0.000 |    654,976 |    676 |   0.000000 | DRAW-PATTERN*
     0.000 |      0.000 |  1,113,040 |  1,371 |   0.000000 | CLIM-CLX::MEDIUM-GCONTEXT
---------------------------------------------------------
     0.153 |      0.000 | 16,400,320 | 41,280 |            | Total
```

These times are with XQuartz 2.7.11 . The linux X11 doesn't seem to share this problem.